+++ banyand/stream/query.go
@@ -19,10 +19,18 @@ package stream
 
 import (
 	"container/heap"
+	"context"
+	"errors"
+	"fmt"
+	"sort"
+
+	"go.uber.org/multierr"
 
 	"github.com/apache/skywalking-banyandb/api/common"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
+	"github.com/apache/skywalking-banyandb/banyand/internal/storage"
 	"github.com/apache/skywalking-banyandb/pkg/convert"
+	itersort "github.com/apache/skywalking-banyandb/pkg/iter/sort"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
 )
@@ -142,7 +150,7 @@ type queryResult struct {
 	ascTS      bool
 }
 
-func (qr *queryResult) Pull() *pbv1.Result {
+func (qr *queryResult) Pull() *pbv1.StreamResult {
 	if !qr.loaded {
 		if len(qr.data) == 0 {
 			return nil
@@ -166,7 +174,7 @@ func (qr *queryResult) Pull() *pbv1.Result {
 		return nil
 	}
 	if len(qr.data) == 1 {
-		r := &pbv1.Result{}
+		r := &pbv1.StreamResult{}
 		bc := qr.data[0]
 		bc.copyAllTo(r, qr.orderByTimestampDesc())
 		qr.data = qr.data[:0]
@@ -225,12 +233,12 @@ func (qr *queryResult) orderByTimestampDesc() bool {
 	return qr.orderByTS && !qr.ascTS
 }
 
-func (qr *queryResult) merge() *pbv1.Result {
+func (qr *queryResult) merge() *pbv1.StreamResult {
 	step := 1
 	if qr.orderByTimestampDesc() {
 		step = -1
 	}
-	result := &pbv1.Result{}
+	result := &pbv1.StreamResult{}
 
 	for qr.Len() > 0 {
 		topBC := qr.data[0]
@@ -254,4 +262,215 @@ func (qr *queryResult) merge() *pbv1.Result {
 	}
 
 	return result
+}
+
+func (s *stream) Filter(ctx context.Context, sfo pbv1.StreamFilterOptions) (sfr pbv1.StreamFilterResult, err error) {
+	if sfo.TimeRange == nil || sfo.Entities == nil {
+		return nil, errors.New("invalid query options: timeRange and series are required")
+	}
+	if len(sfo.TagProjection) == 0 {
+		return nil, errors.New("invalid query options: tagProjection is required")
+	}
+	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
+	tabWrappers := tsdb.SelectTSTables(*sfo.TimeRange)
+	sort.Slice(tabWrappers, func(i, j int) bool {
+		return tabWrappers[i].GetTimeRange().Start.Before(tabWrappers[j].GetTimeRange().Start)
+	})
+	defer func() {
+		for i := range tabWrappers {
+			tabWrappers[i].DecRef()
+		}
+	}()
+
+	var seriesList pbv1.SeriesList
+	for _, entity := range sfo.Entities {
+		sl, lookupErr := tsdb.Lookup(ctx, &pbv1.Series{Subject: sfo.Name, EntityValues: entity})
+		if lookupErr != nil {
+			return nil, lookupErr
+		}
+		seriesList = seriesList.Merge(sl)
+	}
+
+	ces := newColumnElements()
+	for _, tw := range tabWrappers {
+		if len(ces.timestamp) >= sfo.MaxElementSize {
+			break
+		}
+		index := tw.Table().Index()
+		erl, err := index.Search(ctx, seriesList, sfo.Filter)
+		if err != nil {
+			return nil, err
+		}
+		if len(ces.timestamp)+len(erl) > sfo.MaxElementSize {
+			erl = erl[:sfo.MaxElementSize-len(ces.timestamp)]
+		}
+		for _, er := range erl {
+			e, err := tw.Table().getElement(er.seriesID, common.ItemID(er.timestamp), sfo.TagProjection)
+			if err != nil {
+				return nil, err
+			}
+			ces.BuildFromElement(e, sfo.TagProjection)
+		}
+	}
+	return ces, nil
+}
+
+func (s *stream) Sort(ctx context.Context, sso pbv1.StreamSortOptions) (ssr pbv1.StreamSortResult, err error) {
+	if sso.TimeRange == nil || sso.Entities == nil {
+		return nil, errors.New("invalid query options: timeRange and series are required")
+	}
+	if len(sso.TagProjection) == 0 {
+		return nil, errors.New("invalid query options: tagProjection is required")
+	}
+	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
+	tabWrappers := tsdb.SelectTSTables(*sso.TimeRange)
+	defer func() {
+		for i := range tabWrappers {
+			tabWrappers[i].DecRef()
+		}
+	}()
+
+	var seriesList pbv1.SeriesList
+	for _, entity := range sso.Entities {
+		sl, lookupErr := tsdb.Lookup(ctx, &pbv1.Series{Subject: sso.Name, EntityValues: entity})
+		if lookupErr != nil {
+			return nil, lookupErr
+		}
+		seriesList = seriesList.Merge(sl)
+	}
+
+	var iters []*searcherIterator
+	for _, series := range seriesList {
+		seriesSpan := newSeriesSpan(ctx, sso.TimeRange, tabWrappers, series.ID)
+		seekerBuilder := seriesSpan.Build()
+		seekerBuilder.enhance(sso.Filter, sso.Order.Index, sso.Order.Sort, sso.TagProjection)
+		seriesIters, buildErr := seekerBuilder.buildSeriesByIndex()
+		if err != nil {
+			return nil, buildErr
+		}
+		if len(seriesIters) > 0 {
+			iters = append(iters, seriesIters...)
+		}
+	}
+
+	if len(iters) == 0 {
+		return ssr, nil
+	}
+
+	it := newItemIter(iters, sso.Order.Sort)
+	defer func() {
+		err = multierr.Append(err, it.Close())
+	}()
+
+	ces := newColumnElements()
+	for it.Next() {
+		nextItem := it.Val()
+		e, err := nextItem.Element()
+		if err != nil {
+			return nil, err
+		}
+		ces.BuildFromElement(e, sso.TagProjection)
+		if len(ces.timestamp) >= sso.MaxElementSize {
+			break
+		}
+	}
+	return ces, nil
+}
+
+func (s *stream) Query(ctx context.Context, sqo pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error) {
+	if sqo.TimeRange == nil || sqo.Entity == nil {
+		return nil, errors.New("invalid query options: timeRange and series are required")
+	}
+	if len(sqo.TagProjection) == 0 {
+		return nil, errors.New("invalid query options: tagProjection is required")
+	}
+	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
+	tabWrappers := tsdb.SelectTSTables(*sqo.TimeRange)
+	defer func() {
+		for i := range tabWrappers {
+			tabWrappers[i].DecRef()
+		}
+	}()
+	sl, err := tsdb.Lookup(ctx, &pbv1.Series{Subject: sqo.Name, EntityValues: sqo.Entity})
+	if err != nil {
+		return nil, err
+	}
+
+	var result queryResult
+	if len(sl) < 1 {
+		return &result, nil
+	}
+	var sids []common.SeriesID
+	for i := range sl {
+		sids = append(sids, sl[i].ID)
+	}
+	var parts []*part
+	qo := queryOptions{
+		StreamQueryOptions: sqo,
+		minTimestamp:       sqo.TimeRange.Start.UnixNano(),
+		maxTimestamp:       sqo.TimeRange.End.UnixNano(),
+		includeMin:         sqo.TimeRange.IncludeStart,
+		includeMax:         sqo.TimeRange.IncludeEnd,
+	}
+	var n int
+	for i := range tabWrappers {
+		s := tabWrappers[i].Table().currentSnapshot()
+		if s == nil {
+			continue
+		}
+		parts, n = s.getParts(parts, qo.minTimestamp, qo.maxTimestamp)
+		if n < 1 {
+			s.decRef()
+			continue
+		}
+		result.snapshots = append(result.snapshots, s)
+	}
+	// TODO: cache tstIter
+	var tstIter tstIter
+	originalSids := make([]common.SeriesID, len(sids))
+	copy(originalSids, sids)
+	sort.Slice(sids, func(i, j int) bool { return sids[i] < sids[j] })
+	tstIter.init(parts, sids, qo.minTimestamp, qo.maxTimestamp)
+	if tstIter.Error() != nil {
+		return nil, fmt.Errorf("cannot init tstIter: %w", tstIter.Error())
+	}
+	for tstIter.nextBlock() {
+		bc := generateBlockCursor()
+		p := tstIter.piHeap[0]
+		bc.init(p.p, p.curBlock, qo)
+		result.data = append(result.data, bc)
+	}
+	if tstIter.Error() != nil {
+		return nil, fmt.Errorf("cannot iterate tstIter: %w", tstIter.Error())
+	}
+	if sqo.Order == nil {
+		result.orderByTS = true
+		result.ascTS = true
+		return &result, nil
+	}
+	if sqo.Order.Index == nil {
+		result.orderByTS = true
+		if sqo.Order.Sort == modelv1.Sort_SORT_ASC || sqo.Order.Sort == modelv1.Sort_SORT_UNSPECIFIED {
+			result.ascTS = true
+		}
+		return &result, nil
+	}
+
+	result.sidToIndex = make(map[common.SeriesID]int)
+	for i, si := range originalSids {
+		result.sidToIndex[si] = i
+	}
+	return &result, nil
+}
+
+// newItemIter returns a ItemIterator which mergers several tsdb.Iterator by input sorting order.
+func newItemIter(iters []*searcherIterator, s modelv1.Sort) itersort.Iterator[item] {
+	var ii []itersort.Iterator[item]
+	for _, iter := range iters {
+		ii = append(ii, iter)
+	}
+	if s == modelv1.Sort_SORT_DESC {
+		return itersort.NewItemIter[item](ii, true)
+	}
+	return itersort.NewItemIter[item](ii, false)
 }
