+++ banyand/stream/stream.go
@@ -21,21 +21,11 @@ package stream
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"io"
-	"sort"
 	"time"
 
-	"go.uber.org/multierr"
-
-	"github.com/apache/skywalking-banyandb/api/common"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
-	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
-	"github.com/apache/skywalking-banyandb/banyand/internal/storage"
-	"github.com/apache/skywalking-banyandb/banyand/tsdb"
-	itersort "github.com/apache/skywalking-banyandb/pkg/iter/sort"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 	"github.com/apache/skywalking-banyandb/pkg/partition"
 	pbv1 "github.com/apache/skywalking-banyandb/pkg/pb/v1"
@@ -55,7 +45,8 @@ const (
 )
 
 type option struct {
-	flushTimeout time.Duration
+	flushTimeout             time.Duration
+	elementIndexFlushTimeout time.Duration
 }
 
 // Query allow to retrieve elements in a series of streams.
@@ -102,225 +93,6 @@ func (s *stream) Close() error {
 func (s *stream) parseSpec() {
 	s.name, s.group = s.schema.GetMetadata().GetName(), s.schema.GetMetadata().GetGroup()
 	s.indexRuleLocators = partition.ParseIndexRuleLocators(s.schema.GetTagFamilies(), s.indexRules)
-}
-
-func (s *stream) ParseElementIDDeprecated(item tsdb.Item) (string, error) {
-	rawBytes, err := item.Val()
-	if err != nil {
-		return "", err
-	}
-	return string(rawBytes), nil
-}
-
-// NewItemIter returns a ItemIterator which mergers several tsdb.Iterator by input sorting order.
-func NewItemIter(iters []*searcherIterator, s modelv1.Sort) itersort.Iterator[item] {
-	var ii []itersort.Iterator[item]
-	for _, iter := range iters {
-		ii = append(ii, iter)
-	}
-	if s == modelv1.Sort_SORT_DESC {
-		return itersort.NewItemIter[item](ii, true)
-	}
-	return itersort.NewItemIter[item](ii, false)
-}
-
-func (s *stream) Filter(ctx context.Context, sfo pbv1.StreamFilterOptions) (sfr pbv1.StreamFilterResult, err error) {
-	if sfo.TimeRange == nil || sfo.Entities == nil {
-		return nil, errors.New("invalid query options: timeRange and series are required")
-	}
-	if len(sfo.TagProjection) == 0 {
-		return nil, errors.New("invalid query options: tagProjection is required")
-	}
-	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
-	tabWrappers := tsdb.SelectTSTables(*sfo.TimeRange)
-	sort.Slice(tabWrappers, func(i, j int) bool {
-		return tabWrappers[i].GetTimeRange().Start.Before(tabWrappers[j].GetTimeRange().Start)
-	})
-	defer func() {
-		for i := range tabWrappers {
-			tabWrappers[i].DecRef()
-		}
-	}()
-
-	var seriesList pbv1.SeriesList
-	for _, entity := range sfo.Entities {
-		sl, lookupErr := tsdb.Lookup(ctx, &pbv1.Series{Subject: sfo.Name, EntityValues: entity})
-		if lookupErr != nil {
-			return nil, lookupErr
-		}
-		seriesList = seriesList.Merge(sl)
-	}
-
-	ces := newColumnElements()
-	for _, tw := range tabWrappers {
-		if len(ces.timestamp) >= sfo.MaxElementSize {
-			break
-		}
-		index := tw.Table().Index()
-		erl, err := index.Search(ctx, seriesList, sfo.Filter)
-		if err != nil {
-			return nil, err
-		}
-		if len(ces.timestamp)+len(erl) > sfo.MaxElementSize {
-			erl = erl[:sfo.MaxElementSize-len(ces.timestamp)]
-		}
-		for _, er := range erl {
-			e, err := tw.Table().getElement(er.seriesID, common.ItemID(er.timestamp), sfo.TagProjection)
-			if err != nil {
-				return nil, err
-			}
-			ces.BuildFromElement(e, sfo.TagProjection)
-		}
-	}
-	return ces, nil
-}
-
-func (s *stream) Sort(ctx context.Context, sso pbv1.StreamSortOptions) (ssr pbv1.StreamSortResult, err error) {
-	if sso.TimeRange == nil || sso.Entities == nil {
-		return nil, errors.New("invalid query options: timeRange and series are required")
-	}
-	if len(sso.TagProjection) == 0 {
-		return nil, errors.New("invalid query options: tagProjection is required")
-	}
-	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
-	tabWrappers := tsdb.SelectTSTables(*sso.TimeRange)
-	defer func() {
-		for i := range tabWrappers {
-			tabWrappers[i].DecRef()
-		}
-	}()
-
-	var seriesList pbv1.SeriesList
-	for _, entity := range sso.Entities {
-		sl, lookupErr := tsdb.Lookup(ctx, &pbv1.Series{Subject: sso.Name, EntityValues: entity})
-		if lookupErr != nil {
-			return nil, lookupErr
-		}
-		seriesList = seriesList.Merge(sl)
-	}
-
-	var iters []*searcherIterator
-	for _, series := range seriesList {
-		seriesSpan := newSeriesSpan(ctx, sso.TimeRange, tabWrappers, series.ID)
-		seekerBuilder := seriesSpan.Build()
-		seekerBuilder.enhance(sso.Filter, sso.Order.Index, sso.Order.Sort, sso.TagProjection)
-		seriesIters, buildErr := seekerBuilder.buildSeriesByIndex()
-		if err != nil {
-			return nil, buildErr
-		}
-		if len(seriesIters) > 0 {
-			iters = append(iters, seriesIters...)
-		}
-	}
-
-	if len(iters) == 0 {
-		return ssr, nil
-	}
-
-	it := NewItemIter(iters, sso.Order.Sort)
-	defer func() {
-		err = multierr.Append(err, it.Close())
-	}()
-
-	ces := newColumnElements()
-	for it.Next() {
-		nextItem := it.Val()
-		e, err := nextItem.Element()
-		if err != nil {
-			return nil, err
-		}
-		ces.BuildFromElement(e, sso.TagProjection)
-		if len(ces.timestamp) >= sso.MaxElementSize {
-			break
-		}
-	}
-	return ces, nil
-}
-
-func (s *stream) Query(ctx context.Context, sqo pbv1.StreamQueryOptions) (pbv1.StreamQueryResult, error) {
-	if sqo.TimeRange == nil || sqo.Entity == nil {
-		return nil, errors.New("invalid query options: timeRange and series are required")
-	}
-	if len(sqo.TagProjection) == 0 {
-		return nil, errors.New("invalid query options: tagProjection is required")
-	}
-	tsdb := s.databaseSupplier.SupplyTSDB().(storage.TSDB[*tsTable, option])
-	tabWrappers := tsdb.SelectTSTables(*sqo.TimeRange)
-	defer func() {
-		for i := range tabWrappers {
-			tabWrappers[i].DecRef()
-		}
-	}()
-	sl, err := tsdb.Lookup(ctx, &pbv1.Series{Subject: sqo.Name, EntityValues: sqo.Entity})
-	if err != nil {
-		return nil, err
-	}
-
-	var result queryResult
-	if len(sl) < 1 {
-		return &result, nil
-	}
-	var sids []common.SeriesID
-	for i := range sl {
-		sids = append(sids, sl[i].ID)
-	}
-	var parts []*part
-	qo := queryOptions{
-		StreamQueryOptions: sqo,
-		minTimestamp:       sqo.TimeRange.Start.UnixNano(),
-		maxTimestamp:       sqo.TimeRange.End.UnixNano(),
-		includeMin:         sqo.TimeRange.IncludeStart,
-		includeMax:         sqo.TimeRange.IncludeEnd,
-	}
-	var n int
-	for i := range tabWrappers {
-		s := tabWrappers[i].Table().currentSnapshot()
-		if s == nil {
-			continue
-		}
-		parts, n = s.getParts(parts, qo.minTimestamp, qo.maxTimestamp)
-		if n < 1 {
-			s.decRef()
-			continue
-		}
-		result.snapshots = append(result.snapshots, s)
-	}
-	// TODO: cache tstIter
-	var tstIter tstIter
-	originalSids := make([]common.SeriesID, len(sids))
-	copy(originalSids, sids)
-	sort.Slice(sids, func(i, j int) bool { return sids[i] < sids[j] })
-	tstIter.init(parts, sids, qo.minTimestamp, qo.maxTimestamp)
-	if tstIter.Error() != nil {
-		return nil, fmt.Errorf("cannot init tstIter: %w", tstIter.Error())
-	}
-	for tstIter.nextBlock() {
-		bc := generateBlockCursor()
-		p := tstIter.piHeap[0]
-		bc.init(p.p, p.curBlock, qo)
-		result.data = append(result.data, bc)
-	}
-	if tstIter.Error() != nil {
-		return nil, fmt.Errorf("cannot iterate tstIter: %w", tstIter.Error())
-	}
-	if sqo.Order == nil {
-		result.orderByTS = true
-		result.ascTS = true
-		return &result, nil
-	}
-	if sqo.Order.Index == nil {
-		result.orderByTS = true
-		if sqo.Order.Sort == modelv1.Sort_SORT_ASC || sqo.Order.Sort == modelv1.Sort_SORT_UNSPECIFIED {
-			result.ascTS = true
-		}
-		return &result, nil
-	}
-
-	result.sidToIndex = make(map[common.SeriesID]int)
-	for i, si := range originalSids {
-		result.sidToIndex[si] = i
-	}
-	return &result, nil
 }
 
 type streamSpec struct {
