+++ banyand/stream/tstable_test.go
@@ -90,7 +90,7 @@ func Test_tsTable_mustAddDataPoints(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			tmpPath, _ := test.Space(require.New(t))
-			index, _ := newElementIndex(context.TODO(), tmpPath)
+			index, _ := newElementIndex(context.TODO(), tmpPath, 0)
 			tst := &tsTable{
 				index:         index,
 				loopCloser:    run.NewCloser(2),
@@ -224,7 +224,7 @@ func Test_tstIter(t *testing.T) {
 		for _, tt := range tests {
 			t.Run(tt.name, func(t *testing.T) {
 				tmpPath, defFn := test.Space(require.New(t))
-				index, _ := newElementIndex(context.TODO(), tmpPath)
+				index, _ := newElementIndex(context.TODO(), tmpPath, 0)
 				defer defFn()
 				tst := &tsTable{
 					index:         index,
@@ -301,7 +301,7 @@ func Test_tstIter(t *testing.T) {
 					defer defFn()
 
 					tst, err := newTSTable(fileSystem, tmpPath, common.Position{},
-						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: 0})
+						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: 0, elementIndexFlushTimeout: 0})
 					require.NoError(t, err)
 					for i, es := range tt.esList {
 						tst.mustAddElements(es)
@@ -349,7 +349,7 @@ func Test_tstIter(t *testing.T) {
 					defer defFn()
 
 					tst, err := newTSTable(fileSystem, tmpPath, common.Position{},
-						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: defaultFlushTimeout})
+						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: defaultFlushTimeout, elementIndexFlushTimeout: defaultFlushTimeout})
 					require.NoError(t, err)
 					for _, es := range tt.esList {
 						tst.mustAddElements(es)
@@ -376,7 +376,7 @@ func Test_tstIter(t *testing.T) {
 					}
 					// reopen the table
 					tst, err = newTSTable(fileSystem, tmpPath, common.Position{},
-						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: defaultFlushTimeout})
+						logger.GetLogger("test"), timestamp.TimeRange{}, option{flushTimeout: defaultFlushTimeout, elementIndexFlushTimeout: defaultFlushTimeout})
 					require.NoError(t, err)
 					verify(t, tt, tst)
 				})
