+++ pkg/index/index.go
@@ -249,11 +249,6 @@ type SeriesStore interface {
 	SearchWildcard([]byte) ([]Series, error)
 }
 
-// ElementStore is an abstract of a element repository.
-type ElementStore interface {
-	Store
-}
-
 // GetSearcher returns a searcher associated with input index rule type.
 type GetSearcher func(location databasev1.IndexRule_Type) (Searcher, error)
